@@ -3,34 +3,41 @@
   "version": "0.5.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^5.9.0",
-    "@reduxjs/toolkit": "^1.6.1",
-    "@tailwindcss/postcss7-compat": "^2.1.4",
-    "autoprefixer": "^9.8.6",
-    "axios": "^0.21.4",
+    "@reduxjs/toolkit": "^1.7.1",
+    "axios": "^0.24.0",
     "framer-motion": "^4.1.17",
     "moment": "^2.29.1",
     "nprogress": "^0.2.0",
-    "postcss": "^7.0.35",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-query": "^3.25.1",
-    "react-redux": "^7.2.5",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "socket.io-client": "^4.2.0",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.1.4",
+    "react-query": "^3.34.5",
+    "react-redux": "^7.2.6",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "5.0.0",
+    "socket.io-client": "^4.4.0",
     "tailwindcss-classnames": "^2.2.3"
   },
   "scripts": {
-    "start": "cross-env PORT=5000 craco start",
+    "start": "cross-env PORT=5000 react-scripts start",
     "dev": "concurrently \"npm run server\" \"npm run client\"",
-    "build": "set GENERATE_SOURCEMAP=false&& craco build",
-    "test": "craco test",
-    "eject": "react-scripts eject"
+    "build": "set GENERATE_SOURCEMAP=false&& react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "storybook": "start-storybook -p 6006 -s public",
+    "build-storybook": "build-storybook -s public"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": "react-app",
+    "overrides": [
+      {
+        "files": [
+          "**/*.stories.*"
+        ],
+        "rules": {
+          "import/no-anonymous-default-export": "off"
+        }
+      }
+    ]
   },
   "browserslist": {
     "production": [
@@ -45,11 +52,21 @@
     ]
   },
   "devDependencies": {
+    "@storybook/addon-actions": "^6.4.9",
+    "@storybook/addon-essentials": "^6.4.9",
+    "@storybook/addon-links": "^6.4.9",
+    "@storybook/node-logger": "^6.4.9",
+    "@storybook/preset-create-react-app": "^3.2.0",
+    "@storybook/react": "^6.4.9",
     "@types/jest": "^27.0.1",
     "@types/nprogress": "^0.2.0",
-    "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.3.0",
-    "typescript": "^4.4.3"
+    "@types/react": "^17.0.37",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.2",
+    "@welldone-software/why-did-you-render": "^6.2.3",
+    "autoprefixer": "^10.4.0",
+    "postcss": "^8.4.5",
+    "tailwindcss": "^3.0.7",
+    "typescript": "^4.5.4"
   }
 }
