@@ -1,5 +1,4 @@
 import React, { useEffect } from 'react'
-import { History } from 'history'
 import axios from 'axios'
 import { useDispatch, useSelector } from 'react-redux'
 
@@ -7,6 +6,7 @@ import { LightButton, DarkButton } from '../../tools/Inputs'
 import { RootState } from '../../store/store'
 import { loginFail, logoutSuccess } from '../../store/features/userSlice'
 import { checkLogged } from '../../actions/userActions'
+import { useNavigate } from 'react-router-dom'
 
 export const logoutAction = (dispatch: React.Dispatch<any>) => {
   axios
@@ -19,21 +19,21 @@ export const logoutAction = (dispatch: React.Dispatch<any>) => {
     })
 }
 
-const Logout = (props: { history: History }) => {
+const Logout = () => {
   const dispatch = useDispatch()
   // 通过选择器获取state
   const logged = useSelector((state: RootState) => state.user.logged)
+  const navigate = useNavigate()
 
   useEffect(() => {
     ;(async () => {
       const data = await checkLogged(dispatch)
-      if (data?.code !== 0 || logged === false)
-        props.history.push('/user/login')
+      if (data?.code !== 0 || logged === false) navigate('/user/login')
     })()
-  }, [logged, dispatch, props.history])
+  }, [logged, dispatch, navigate])
 
   const cancel = () => {
-    props.history.push('/')
+    navigate('/')
   }
 
   return (
